# backend/disney_OpenAPI.yaml
openapi: 3.0.0
info:
  title: API Disney
  description: API para la gestión de personajes, géneros, usuarios y audiovisuales.
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /characters:
    get:
      summary: Obtener todos los personajes o buscar por filtros
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filtrar por nombre de personaje
        - name: age
          in: query
          required: false
          schema:
            type: integer
          description: Filtrar por edad del personaje
        - name: weight
          in: query
          required: false
          schema:
            type: number
          description: Filtrar por peso del personaje
        - name: audiovisualId
          in: query
          required: false
          schema:
            type: integer
          description: Filtrar por ID de audiovisual relacionado
      responses:
        '200':
          description: Lista de personajes, con o sin filtros aplicados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'
      security:
        - jwtAuth: []
    post:
      summary: Crear un personaje
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                edad:
                  type: integer
                peso:
                  type: number
                historia:
                  type: string
                imagen:
                  type: string
                  format: binary
      responses:
        '201':
          description: Personaje creado
      security:
        - jwtAuth: []
  /characters/{id}:
    get:
      summary: Obtener personaje por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del personaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
      security:
        - jwtAuth: []
    delete:
      summary: Eliminar un personaje por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Personaje eliminado
      security:
        - jwtAuth: []
    put:
      summary: Modificar parcialmente un personaje
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                edad:
                  type: integer
                peso:
                  type: number
                historia:
                  type: string
                imagen:
                  type: string
                  format: binary
      responses:
        '200':
          description: Personaje modificado parcialmente
      security:
        - jwtAuth: []
  /genre:
    get:
      summary: Obtener todos los géneros
      responses:
        '200':
          description: Lista de géneros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
      security:
        - jwtAuth: []
    post:
      summary: Crear un género
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                imagen:
                  type: string
                  format: binary
      responses:
        '201':
          description: Género creado
      security:
        - jwtAuth: []
  /genre/{id}:
    put:
      summary: Modificar un género
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                imagen:
                  type: string
                  format: binary
      responses:
        '200':
          description: Género modificado
      security:
        - jwtAuth: []
    delete:
      summary: Eliminar un género
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Género eliminado
      security:
        - jwtAuth: []
    patch:
      summary: Modificar parcialmente un género
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                imagen:
                  type: string
                  format: binary
      responses:
        '200':
          description: Género modificado parcialmente
      security:
        - jwtAuth: []
...
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Character:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        edad:
          type: integer
        peso:
          type: number
        historia:
          type: string
        imagen:
          type: string
    Genre:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        imagen:
          type: string
...